DefaultVisualizer writes graphs in both Dot and DGML format.
